feat!: validate Avro schemas included in AsyncAPI docs
This commit adds validation to the avro-schema-parser so that
invalid Avro schemas will be rejected.

I had to fix a few invalid Avro schemas in the tests (which were
using the type 'uuid' which isn't in the Avro spec) so this should
potentially be considered a major breaking change to the parser.

Closes: #37

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>